Fix log append to account for container engine, promote fixture
