+++ tests/integration/actions/doc/test_stdout.py
@@ -6,7 +6,7 @@ import pytest
 
 from .base import BaseClass
 
-from ..._common import container_runtime_or_fail
+from ...._common import container_runtime_or_fail
 
 
 # ansible-doc help with EE
