+++ tests/integration/actions/ee_details/base.py
@@ -13,10 +13,11 @@ from typing import Union
 
 import pytest
 
-from ..._common import container_runtime_or_fail
 from ..._common import fixture_path_from_request
 from ..._common import update_fixtures
 from ..._tmux_session import TmuxSession
+
+from ...._common import container_runtime_or_fail
 from ....defaults import FIXTURES_DIR
