+++ tests/integration/actions/stdout/test_welcome_interactive_ee.py
@@ -5,7 +5,7 @@ import pytest
 from .base import BaseClass
 from .base import ANSIBLE_PLAYBOOK
 
-from ..._common import container_runtime_or_fail
+from ...._common import container_runtime_or_fail
 
 CLI = "ansible-navigator" " --execution-environment true --ce " + container_runtime_or_fail()
