+++ tests/unit/test_cli.py
@@ -8,8 +8,6 @@ import pytest
 from ansible_navigator.cli import parse_and_update
 from ansible_navigator.cli import NavigatorConfiguration
 
-from ansible_navigator.configuration_subsystem import Constants as C
-
 from ..defaults import FIXTURES_DIR
